Change matrix representation to use vector of rows format
As a result of changing the matrix representation format, the syntax for
accessing elements in a matrix has changed from (aref matrix i j) to
(aref (aref matrix i) j) - a little more verbose. However, the new
representation allows for easily shuffling the rows of the matrix and also for
easily extracting a row which corresponding to a training sample.