Add several updates and fix bugs.
This commits corrects the equation for the cross entropy used to compute the
cost. A function for computing accuracy was also added.

This commit also fixes that results because of the way the data is
represented. The code runs now but it's very slow. We will plan to optimize it
later on